\name{plot_lefse_LDA_graph}
\alias{plot_lefse_LDA_graph}

\title{
Create LEfSe plot for taxa
}
\description{
Creates an adaptation of a LEfSe plot (https://galaxyproject.org/learn/visualization/custom/lefse/) of taxa that are considered significantlly different by LEfSe analysis.
}
\usage{
plot_lefse_LDA_graph()
}

\arguments{
None
}
\details{

}
\value{
None
}
\references{

}
\author{
Roshonda Barner Jones
}
\note{

}

\seealso{

}
\examples{
 plot_lefse_LDA_graph()
}

\keyword{ ~aplot }
