\name{plot_mds_coloredByRapidGrowth}
\alias{plot_mds_coloredByRapidGrowth}

\title{
Create 2-dimensional PCA plot by rapid growth status
}
\description{
Create an PCA plot with points colored by rapid growth status. The MDS axes were using a distance metric deicode which is a robust form of Aitchison distance (doi:10.1128/mSystems.00016-19).
}
\usage{
plot_mds_coloredByRapidGrowth()
}

\arguments{
None
}
\details{

}
\value{
None
}
\references{

}
\author{
Roshonda Barner Jones
}
\note{

}


\seealso{

}
\examples{
plot_mds_coloredByRapidGrowth()
}

\keyword{ ~aplot }
