\name{plot_phylum_relativeAbundances_barplot}
\alias{plot_phylum_relativeAbundances_barplot}

\title{
Plot phylum-level taxa by rapid growth status.
}
\description{
Create a barplot with the relative abundances of taxa from gut microbiome of 1-month old infants. The barplot is separated by rapid growth status.
}
\usage{
plot_phylum_relativeAbundances_barplot()
}

\arguments{
None
}
\details{

}
\value{
None
}
\references{

}
\author{
Roshonda Barner Jones
}

\note{
}

\seealso{
}
\examples{
plot_phylum_relativeAbundances_barplot()
}

\keyword{ ~aplot }

